BIG REFACTOR: Fix bug.
CONTEXT: I am trying to check if the current implementation instantiates
'cpu' structs, and also wanted to know if the cpu_run() runs. The cpu_run()
is just a sketch, there are at this commit a lot of functionalities pending to implement,
but at least it is supposed to run an infinite loop that prints some 'default'
statement.

THE BUG: Compiled code always returns 'Segmentation fault'
I really do not understand where this comes from, so I added some 'printf'
statements in the 'ls8.c->main()' with the aim of getting a signal from where
comes the error.

After compiling again I noticed that the no one of the 'printf' printed out.
So I decided to refactor the 'struct CPU'. After that the new compilation
runs.